Use static class methods instead of global functions
Files defining functions cannot be autoloaded and need to be required
proactively, leading to fatals if the library is known to multiple
autoloaders (something that is likely to happen with our Composer
setup).

As a temporary measure put the functions behind a duplication check
(which will probably have a bad effect on opcaches); as a long-term
solution deprecate them and move the logic into static class methods.

Bug: T186182
Change-Id: I7801d6f6d58b2e6fceeb62ba8316a83bfb919a0a